Reengineering branch: better software architecture design (still in progress)
