+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".MainActivity"
+            android:name=".activity.MainActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
@@ -24,8 +24,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".UserWishlistActivity" />
-        <activity android:name=".AddItemActivity"></activity>
+        <activity android:name=".activity.UserWishlistActivity" />
+        <activity android:name=".activity.AddItemActivity"></activity>
     </application>
 
 </manifest>
