+++ app/src/test/java/com/peppedalterio/whatsyourwish/AddItemValidateTest.java
@@ -1,4 +1,4 @@
-package com.peppedalterio.whatsyourwish;
+package com.peppedalterio.whatsyourwish.activity;
 
 import org.junit.jupiter.api.*;
